requirements.txt
# requirements.txt

requests
numpy
faiss-cpu
pdfplumber